+++ src/pages/Hall/index.js
@@ -18,6 +18,7 @@ import Sides from '../../components/All-day/Sides';
 import InputSelect from '../../components/InputSelect';
 
 import Popup from '../../components/Popup';
+import DateTime from '../../components/DateTime'
 
 export default function Hall() {
 	
@@ -276,60 +277,12 @@ export default function Hall() {
 	}, []);
 	console.log(allProducts);
 
-	const newDate = new Date();
-	const dayName = [
-		'Domingo',
-		'Segunda-feira',
-		'Terça-feira',
-		'Quarta-feira',
-		'Quinta-feira',
-		'Sexta',
-		'Sábado',
-	];
-	const week = dayName[newDate.getDay()];
-	const day = String(newDate.getDate()).padStart(2, '0');
-	const monthName = [
-		'jan,',
-		'fev,',
-		'mar.',
-		'abr.',
-		'maio',
-		'jun.',
-		'jul.',
-		'ago.',
-		'set.',
-		'out.',
-		'nov.',
-		'dez.',
-	];
-	const month = monthName[newDate.getMonth()];
-	const year = newDate.getFullYear();
-
-	const [currentHour, serCurrentHour] = useState(new Date());
-
-	function refreshClock() {
-		serCurrentHour(new Date());
-	}
-	useEffect(() => {
-		const timerId = setInterval(refreshClock, 1000);
-		return function cleanup() {
-			clearInterval(timerId);
-		};
-	}, []);
-
 	return (
 		<div className="pages-container">
 			<header className="title-area">
 				<div>
 					<h1 className="uppercase">Divino Burger</h1>
-					<p>
-						{day} {month} {year}, {week} - 🕐
-						{currentHour.toLocaleTimeString('pt-br', {
-							hour: '2-digit',
-							minute: '2-digit',
-						})}
-						h
-					</p>
+					<DateTime/>
 				</div>
 
 				<ButtonLogout />
